refactor: extract Context into src/snippets/context.ts
